Revert "[TTNN] Introduce PrepareConv2dBias op (#3843)" (#3937)
This reverts commit 4d38040e3f48b2b28153cc3091c35e05d8f346b4.

### Problem description
ResNet fails when the optimizer is enabled because the TTNNOpModel
library does not support the `PrepareConv2dBias` operation. I will
reapply the commit once support is added to the TTNNOpModel library.
