+++ include/ttmlir/Dialect/TTNN/IR/TTNNOps.td
@@ -1234,30 +1234,6 @@ def TTNN_PrepareConv2dWeightsOp : TTNN_Op<"prepare_conv2d_weights"> {
     let hasVerifier = 1;
 }
 
-def TTNN_PrepareConv2dBiasOp : TTNN_Op<"prepare_conv2d_bias"> {
-    let summary = "Prepares conv2d bias so that it can be consumed by the conv2d op.";
-
-    let arguments = (ins AnyRankedTensor:$bias_tensor,
-                         TTNN_MemoryConfigAttr:$input_memory_config,
-                         TTNN_LayoutAttr:$input_tensor_layout,
-                         I32Attr:$in_channels,
-                         I32Attr:$out_channels,
-                         I32Attr:$batch_size,
-                         I32Attr:$input_height,
-                         I32Attr:$input_width,
-                         DenseI32ArrayAttr:$kernel_size,
-                         DenseI32ArrayAttr:$stride,
-                         DenseI32ArrayAttr:$padding,
-                         DenseI32ArrayAttr:$dilation,
-                         I32Attr:$groups,
-                         TTNN_Device:$device,
-                         OptionalAttr<TTNN_Conv2dConfigAttr>:$conv2d_config);
-
-    let results = (outs AnyRankedTensor:$result);
-
-    let hasVerifier = 1;
-}
-
 def TTNN_Conv2dOp : TTNN_Op<"conv2d",
       [DeclareOpInterfaceMethods<TTNN_OpModelInterface, ["getOpConstraints", "getOpRuntime"]>]
       > {
