+++ include/ttmlir/Dialect/TTNN/Transforms/Passes.td
@@ -134,12 +134,12 @@ def TTNNTuplifyTensors: Pass<"ttnn-tuplify-tensors", "::mlir::ModuleOp"> {
   ];
 }
 
-def TTNNPrepareConv2dWeightsAndBias : Pass<"ttnn-prepare-conv2d-weights-and-bias", "::mlir::ModuleOp"> {
-  let summary = "Optimize Conv2d ops by inserting PrepareConv2dWeights and PrepareConv2dBias ops before them.";
+def TTNNPrepareConv2dWeights : Pass<"ttnn-prepare-conv2d-weights", "::mlir::ModuleOp"> {
+  let summary = "Optimize Conv2d ops by inserting PrepareConv2dWeights before them.";
   let description = [{
-    This pass inserts a PrepareConv2dWeights and a PrepareConv2dBias operation before each Conv2d op, which preprocess the weights and bias used by the Conv2d operations.
-    These operations can then be const-evaluated, leading to improved performance.
-    To use this pass, the project must be built with the op model library by setting -DTTMLIR_ENABLE_OPMODEL=ON during the build process.
+    This pass inserts a PrepareConv2dWeights operation before each Conv2d op which preprocess the weights used by Conv2d operations.
+    The PrepareConv2dWeights op can be then const-evaled, leading to improved performance. In order to use this pass,
+    the project must be built with the op model library by setting -DTTMLIR_ENABLE_OPMODEL=ON during the build process.
   }];
 }
