+++ runtime/lib/ttnn/program_executor.cpp
@@ -12,7 +12,6 @@
 #include "operations/context/get_device.h"
 #include "operations/conv/conv2d.h"
 #include "operations/conv/conv_transpose2d.h"
-#include "operations/conv/prepare_conv2d_bias.h"
 #include "operations/conv/prepare_conv2d_weights.h"
 #include "operations/cpu/cpu.h"
 #include "operations/creation/arange.h"
@@ -257,10 +256,6 @@ void ProgramExecutor::runOperation(const ::tt::target::ttnn::Operation *op) {
   }
   case ::tt::target::ttnn::OpType::PrepareConv2dWeightsOp: {
     return operations::conv::run(op->type_as_PrepareConv2dWeightsOp(),
-                                 getContext());
-  }
-  case ::tt::target::ttnn::OpType::PrepareConv2dBiasOp: {
-    return operations::conv::run(op->type_as_PrepareConv2dBiasOp(),
                                  getContext());
   }
   case ::tt::target::ttnn::OpType::Conv2dOp: {
