--- test/ttmlir/EmitC/TTNN/conv/prepare_conv2d_bias_asymmetric_padding.mlir
@@ -1,61 +1,0 @@
-// RUN: ttmlir-opt --ttcore-register-device="system-desc-path=%system_desc_path%" %s > %t.mlir
-// RUN: ttmlir-translate --ttnn-to-flatbuffer %t.mlir > %basename_t.ttnn
-// RUN: ttmlir-opt --ttnn-tuplify-tensors --convert-ttnn-to-emitc %t.mlir > %t2.mlir
-// RUN: ttmlir-translate --mlir-to-cpp %t2.mlir > %basename_t.cpp
-
-#dram = #ttnn.buffer_type<dram>
-#system_memory = #ttnn.buffer_type<system_memory>
-
-#ttnn_layout_bias_bf16 = #ttnn.ttnn_layout<(d0, d1, d2, d3) -> (d0 + d1 + d2, d3), <1x1>, memref<1x128xbf16, #system_memory>>
-#ttnn_layout_prepared_bias_bf16 = #ttnn.ttnn_layout<(d0, d1, d2, d3) -> (d0 * 2112 + d1 * 64 + d2, d3), <8x8, (d0, d1) -> (0, d0, d1)>, memref<9x1x!ttcore.tile<32x32, bf16>, #dram>, <interleaved>>
-#conv2d_config_bf16 = #ttnn.conv2d_config<weights_dtype = bf16>
-
-func.func @prepare_conv2d_bias_asymmetric_padding_bf16(%arg0: tensor<1x1x1x128xbf16, #ttnn_layout_bias_bf16>) -> tensor<1x1x1x128xbf16, #ttnn_layout_prepared_bias_bf16> {
-  %0 = "ttnn.get_device"() <{mesh_offset = #ttnn<mesh_offset 0x0>, mesh_shape = #ttnn<mesh_shape 1x1>}> : () -> !ttnn.device
-  %1 = "ttnn.prepare_conv2d_bias"(%arg0, %0)
-      <{
-        batch_size = 1 : i32,
-        conv2d_config = #conv2d_config_bf16,
-        dilation = array<i32: 1, 1>,
-        groups = 1 : i32,
-        in_channels = 64 : i32,
-        input_height = 32 : i32,
-        input_memory_config = #ttnn.memory_config<#dram, <interleaved>>,
-        input_tensor_layout = #ttnn.layout<tile>,
-        input_width = 32 : i32,
-        kernel_size = array<i32: 3, 3>,
-        out_channels = 128 : i32,
-        padding = array<i32: 1, 2, 3, 4>,
-        stride = array<i32: 1, 1>
-      }> : (tensor<1x1x1x128xbf16, #ttnn_layout_bias_bf16>, !ttnn.device) -> tensor<1x1x1x128xbf16, #ttnn_layout_prepared_bias_bf16>
-
-  "ttnn.deallocate"(%arg0) <{force = false}> : (tensor<1x1x1x128xbf16, #ttnn_layout_bias_bf16>) -> ()
-  return %1 : tensor<1x1x1x128xbf16, #ttnn_layout_prepared_bias_bf16>
-}
-
-#ttnn_layout_bias_f32 = #ttnn.ttnn_layout<(d0, d1, d2, d3) -> (d0 + d1 + d2, d3), <1x1>, memref<1x64xf32, #system_memory>>
-#ttnn_layout_prepared_bias_f32 = #ttnn.ttnn_layout<(d0, d1, d2, d3) -> (d0 * 32 + d1 * 32 + d2, d3), <1x1>, memref<1x2x!ttcore.tile<32x32, f32>, #dram>, <interleaved>>
-#conv2d_config_f32 = #ttnn.conv2d_config<weights_dtype = f32>
-
-func.func @prepare_conv2d_bias_f32(%arg0: tensor<1x1x1x64xf32, #ttnn_layout_bias_f32>) -> tensor<1x1x1x64xf32, #ttnn_layout_prepared_bias_f32> {
-  %0 = "ttnn.get_device"() <{mesh_offset = #ttnn<mesh_offset 0x0>, mesh_shape = #ttnn<mesh_shape 1x1>}> : () -> !ttnn.device
-  %1 = "ttnn.prepare_conv2d_bias"(%arg0, %0)
-      <{
-        batch_size = 1 : i32,
-        conv2d_config = #conv2d_config_f32,
-        dilation = array<i32: 1, 1>,
-        groups = 1 : i32,
-        in_channels = 64 : i32,
-        input_height = 32 : i32,
-        input_memory_config = #ttnn.memory_config<#dram, <interleaved>>,
-        input_tensor_layout = #ttnn.layout<tile>,
-        input_width = 16 : i32,
-        kernel_size = array<i32: 5, 5>,
-        out_channels = 64 : i32,
-        padding = array<i32: 2, 1, 4, 5>,
-        stride = array<i32: 1, 1>
-      }> : (tensor<1x1x1x64xf32, #ttnn_layout_bias_f32>, !ttnn.device) -> tensor<1x1x1x64xf32, #ttnn_layout_prepared_bias_f32>
-
-  "ttnn.deallocate"(%arg0) <{force = false}> : (tensor<1x1x1x64xf32, #ttnn_layout_bias_f32>) -> ()
-  return %1 : tensor<1x1x1x64xf32, #ttnn_layout_prepared_bias_f32>
-}
