+++ test/ttmlir/EmitC/TTNN/conv/prepare_conv2d_weights_asymmetric_padding.mlir
@@ -20,10 +20,9 @@ func.func @prepare_conv2d_weights(%arg0: tensor<16x8x3x3xbf16, #ttnn_layout>) ->
           input_height = 32 : i32,
           input_memory_config = #ttnn.memory_config<#dram, <interleaved>>,
           input_tensor_layout = #ttnn.layout<tile>,
-          input_width = 32 : i32,
-          kernel_size = array<i32: 3, 3>,
+          input_width = 32 : i32, kernel_size = array<i32: 3, 3>,
           out_channels = 16 : i32,
-          padding = array<i32: 0, 2, 1, 3>,
+          padding = array<i32: 0, 1, 2, 3>,
           stride = array<i32: 2, 2>,
           weights_format = "OIHW"
       }> : (tensor<16x8x3x3xbf16, #ttnn_layout>, !ttnn.device) -> tensor<1x1x72x16xbf16, #ttnn_layout1>
